// front-end/src/app/models/stagiaire.model.ts
export class Stagiaire {
   id?: number | null;
    firstname: string = '';
    lastname: string = '';
    adresse: string = '';
    numTel?: number;
    education: string = '';
    email:string = '';
    etablissement: string = '';
    cin?: number ;
    profile: string = '';
  

  }
  